+++ changelog.md
@@ -6,7 +6,8 @@
 - fix for play menu item in filters menu of main dialog being disabled
 - fix audio being not loaded by mpv.net when the avs/vs script is played
 - fix a issue with the custom icon feature
-- x265 --aq-mode update
+- update: x265 3.1+11-de920e0 Wolfberry
+- x265 --aq-mode update, new switches --hme and --hme-search added
 - update: L-Smash r935+31-2019-08-17 HolyWu
   added support for showing indexing progress and cachefile parameter and
   for native high bit depth, avs and vs l-smash are now contained in the same dll,